Move model creation to base package.
We now have multiple suv-packages with model definitions, but creating tables
is easier when done in one place only.
