+++ opengever/globalindex/configure.zcml
@@ -3,7 +3,6 @@
     xmlns:genericsetup="http://namespaces.zope.org/genericsetup"
     xmlns:grok="http://namespaces.zope.org/grok"
     xmlns:i18n="http://namespaces.zope.org/i18n"
-    xmlns:profilehook="http://namespaces.zope.org/profilehook"
     i18n_domain="opengever.globalindex">
 
   <include package=".handlers" />
@@ -19,11 +18,6 @@
       directory="profiles/default"
       description="Installs the opengever.globalindex package"
       provides="Products.GenericSetup.interfaces.EXTENSION"
-      />
-
-  <profilehook:hook
-      profile="opengever.globalindex:default"
-      handler=".hooks.installed"
       />
 
 </configure>
